+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename DoesMemoryMatch method to Compare (#22)
 - Remove DataMatches property from IMemoryComparison interface (#22)
+- Rename SetLogger method to SetApplicationLogger to help avoid confusion (#22)
 - Sheet and range data is now cached to avoid expesive read operations on the provided sheet and range classes (#26)
 
 ### Added